Deploy payments banner in settings page, updated tests
